refactor: redistribute activation related datatypes into eth.rs and utxo.rs
