+++ mm2src/mm2_rpc/src/data/legacy.rs
@@ -3,8 +3,9 @@
 #[path = "legacy/utility.rs"] mod utility;
 #[path = "legacy/wallet.rs"] mod wallet;
 
-pub use activation::{CoinInitResponse, ElectrumProtocol, EnabledCoin, GasStationPricePolicy, GetEnabledResponse,
-                     UtxoMergeParams};
+pub use activation::{eth::GasStationPricePolicy,
+                     utxo::{ElectrumProtocol, UtxoMergeParams},
+                     CoinInitResponse, EnabledCoin, GetEnabledResponse};
 pub use orders::{AggregatedOrderbookEntry, MatchBy, OrderConfirmationsSettings, OrderType, OrderbookRequest,
                  OrderbookResponse, RpcOrderbookEntry, SellBuyRequest, SellBuyResponse, TakerAction,
                  TakerRequestForRpc};
